// src/app/models/encuesta.ts
export class Encuesta {
    usuario: string | undefined | null;
    nombre: string | undefined;
    apellido: string | undefined;
    edad: number | undefined;
    telefono: string | undefined;
    comoNosEncontraste: string | undefined;
    cuantoGusto: number | undefined;
    cualesJugaste = [];
    recomendarias: string | undefined;
}